+++ client/public/locales/en/events.json
@@ -1,0 +1,3 @@
+{
+  
+}
