+++ client/public/locales/ru/events.json
@@ -1,0 +1,3 @@
+{
+  
+}
