+++ client/src/entities/Event/ui/EventForm/EventForm.stories.tsx
@@ -17,7 +17,7 @@ export default meta
 type Story = StoryObj<typeof EventForm>;
 
 const event: Event = {
-    title: 'Тестовый заказ',
+    title: 'Тестовое событие',
     eventDate: '2024-02-12',
     startTime: '12:00',
     endTime: '15:00',
@@ -31,11 +31,37 @@ const event: Event = {
 export const Default: Story = {
     args: {
         data: event,
+        events: [{
+            _id: '643e58deaba80539138865d0',
+            name: 'Учеба',
+            color: '#bde2a8',
+            isDefault: false
+        },
+        {
+            _id: '643e58efaba80539138865d2',
+            name: 'Личное',
+            color: '#ff759f',
+            isDefault: false
+        },
+        {
+            _id: '643e5900aba80539138865d4',
+            name: 'Семья',
+            color: 'gray',
+            isDefault: false
+        },
+        {
+            _id: '6467834500aba6813881d4',
+            name: 'Друзья',
+            color: 'blue',
+            isDefault: false
+        }
+        ],
         errors: {},
         onSave: () => {}
     },
     decorators: [
         WithColorsTheme
+        
     ],
     parameters: {
         docs: {
@@ -43,3 +69,4 @@ export const Default: Story = {
         },
     }
 }
+
