+++ client/src/pages/EventDetailPage/ui/EventDetailPage.stories.tsx
@@ -2,10 +2,11 @@ import type { Meta, StoryObj } from '@storybook/react'
 import 'app/styles/index.scss'
 import { WithColorsTheme } from 'shared/config/storybook/ThemeDecorator/WithColorsTheme'
 import { EventDetailPage } from './EventDetailPage'
+import { eventDetailStoreDecorator } from 'shared/config/storybook/StoreDecorator/EventDetailStoreDecorator'
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta: Meta<typeof EventDetailPage> = {
-    title: 'shared/EventDetailPage',
+    title: 'pages/EventDetailPage',
     component: EventDetailPage,
     argTypes: {
     // backgroundColor: { control: 'color' },
@@ -24,7 +25,8 @@ export const Default: Story = {
                                                                                                                                                           
     },
     decorators: [
-        WithColorsTheme
+        WithColorsTheme,
+        eventDetailStoreDecorator
     ],
     parameters: {
         docs: {
