+++ client/src/shared/config/storybook/StoreDecorator/CalendarStoreDecorator.tsx
@@ -1,25 +1,27 @@
 import { ReducersMapObject } from '@reduxjs/toolkit'
 import { Decorator } from '@storybook/react'
 import { StateSchema, StoreProvider } from 'app/providers/StoreProvider'
-import { clientDetailsReducer } from 'entities/Clients'
-import { stateAllErrors, stateAllIn, stateAllIsLoading } from './state'
+import { calendarPageReducer } from 'pages/CalendarPage'
+import { stateAllErrors, stateAllIn } from './state'
+import { stateLoginFormLoading } from './LoginFormStoreDecorator'
 
 export const defaultAsyncReducers: DeepPartial<ReducersMapObject<StateSchema>> = {
-    clientDetails: clientDetailsReducer
+    CalendarPage: calendarPageReducer,
 }
 
-export const clientDetailsStoreDecorator: Decorator = (Story) => (
+export const CalendarStoreDecorator: Decorator = (Story) => (
     <StoreProvider initialState={stateAllIn} asyncReducers={defaultAsyncReducers}>
         {Story()}
     </StoreProvider>
 )
-export const clientDetailsErrorStoreDecorator: Decorator = (Story) => (
+export const calendarPageIsLoadingStoreDecorator: Decorator = (Story) => (
     <StoreProvider initialState={stateAllErrors} asyncReducers={defaultAsyncReducers}>
         {Story()}
     </StoreProvider>
 )
-export const clientDetailsIsLoadingStoreDecorator: Decorator = (Story) => (
-    <StoreProvider initialState={stateAllIsLoading} asyncReducers={defaultAsyncReducers}>
+
+export const calendarPageErrorStoreDecorator: Decorator = (Story) => (
+    <StoreProvider initialState={stateLoginFormLoading} asyncReducers={defaultAsyncReducers}>
         {Story()}
     </StoreProvider>
 )
