+++ client/src/shared/config/storybook/StoreDecorator/EventDetailStoreDecorator.tsx
@@ -1,25 +1,26 @@
 import { ReducersMapObject } from '@reduxjs/toolkit'
 import { Decorator } from '@storybook/react'
 import { StateSchema, StoreProvider } from 'app/providers/StoreProvider'
-import { clientDetailsReducer } from 'entities/Clients'
 import { stateAllErrors, stateAllIn, stateAllIsLoading } from './state'
+import { eventDetailReducer } from 'pages/EventDetailPage'
 
 export const defaultAsyncReducers: DeepPartial<ReducersMapObject<StateSchema>> = {
-    clientDetails: clientDetailsReducer
+    EventDetailsPage: eventDetailReducer,
 }
 
-export const clientDetailsStoreDecorator: Decorator = (Story) => (
+export const eventDetailStoreDecorator: Decorator = (Story) => (
     <StoreProvider initialState={stateAllIn} asyncReducers={defaultAsyncReducers}>
         {Story()}
     </StoreProvider>
 )
-export const clientDetailsErrorStoreDecorator: Decorator = (Story) => (
-    <StoreProvider initialState={stateAllErrors} asyncReducers={defaultAsyncReducers}>
+export const eventDetailIsLoadingStoreDecorator: Decorator = (Story) => (
+    <StoreProvider initialState={stateAllIsLoading} asyncReducers={defaultAsyncReducers}>
         {Story()}
     </StoreProvider>
 )
-export const clientDetailsIsLoadingStoreDecorator: Decorator = (Story) => (
-    <StoreProvider initialState={stateAllIsLoading} asyncReducers={defaultAsyncReducers}>
+
+export const eventDetailErrorStoreDecorator: Decorator = (Story) => (
+    <StoreProvider initialState={stateAllErrors} asyncReducers={defaultAsyncReducers}>
         {Story()}
     </StoreProvider>
 )
