+++ client/src/shared/config/storybook/StoreDecorator/orderStatusStoreDecorator.tsx
@@ -2,6 +2,7 @@ import { ReducersMapObject } from '@reduxjs/toolkit'
 import { Decorator } from '@storybook/react'
 import { StateSchema, StoreProvider } from 'app/providers/StoreProvider'
 import { orderStatusReducer } from 'entities/OrderStatus'
+import { stateAllErrors, stateAllIn } from './state'
 
 export const defaultAsyncReducers: DeepPartial<ReducersMapObject<StateSchema>> = {
     orderStatuses: orderStatusReducer,
@@ -58,12 +59,12 @@ export const orderStatusIsLoading: DeepPartial<StateSchema> = {
 }
 
 export const orderStatusesStoreDecorator: Decorator = (Story) => (
-    <StoreProvider initialState={orderStatuses} asyncReducers={defaultAsyncReducers}>
+    <StoreProvider initialState={stateAllIn} asyncReducers={defaultAsyncReducers}>
         {Story()}
     </StoreProvider>
 )
 export const orderStatusesErrorStoreDecorator: Decorator = (Story) => (
-    <StoreProvider initialState={orderStatusError} asyncReducers={defaultAsyncReducers}>
+    <StoreProvider initialState={stateAllErrors} asyncReducers={defaultAsyncReducers}>
         {Story()}
     </StoreProvider>
 )
