+++ client/src/shared/config/storybook/StoreDecorator/settingPageStoreDecorator.tsx
@@ -1,83 +1,21 @@
 import { Decorator } from '@storybook/react'
 import { StateSchema, StoreProvider } from 'app/providers/StoreProvider'
 import { ReducersMapObject } from '@reduxjs/toolkit'
-import { Country } from 'entities/Country'
-import { Currency } from 'entities/Currency'
 import { profileReducer } from 'entities/Profile'
 import { orderStatusEditReducer } from 'widgets/OrderStatusEdit'
+import { stateAllIn } from './state'
+import { projectEditReducer } from 'widgets/ProjectEdit'
+import { eventTypesEditReducer } from 'widgets/EventTypesEdit'
 
 export const defaultAsyncReducers: DeepPartial<ReducersMapObject<StateSchema>> = {
     profile: profileReducer,
     OrderStatusEditSchema: orderStatusEditReducer,
+    ProjectEdit: projectEditReducer,
+    eventTypesEditSchema: eventTypesEditReducer
 }
-
-export const state: DeepPartial<StateSchema> = {
-    profile: {
-        readonly: true,
-        isLoading: false,
-        error: undefined,
-        data: {
-            firstname:'Tatiana',
-            lastname:'Gorelova',
-            age:38,
-            currency:Currency.RUB,
-            country:Country.Russia,
-            city:'Zvenigorod',
-            username:'tratata',
-            avatar:'https://avatars.githubusercontent.com/u/107557323?v=4'
-        },
-        form: {
-            firstname:'Tatiana',
-            lastname:'Gorelova',
-            age:38,
-            currency:Currency.RUB,
-            country:Country.Russia,
-            city:'Zvenigorod',
-            username:'tratata',
-            avatar:'https://avatars.githubusercontent.com/u/107557323?v=4'
-        }
-    },
-    OrderStatusEditSchema: {
-        isLoading: false,
-        editStatusId: '64492fd3d206f2a8b5f4298e',
-        data: [{
-            _id: '643e58deaba80539138865d0',
-            name: 'Завершен',
-            color: '#bde2a8',
-            isDefault: true
-        },
-        {
-            _id: '643e58efaba80539138865d2',
-            name: 'Отменен',
-            color: '#ff759f',
-            isDefault: true
-        },
-        {
-            _id: '643e5900aba80539138865d4',
-            name: 'Архив',
-            color: 'gray',
-            isDefault: true
-        },
-        {
-            _id: '6467834500aba6813881d4',
-            name: 'Новый',
-            color: 'blue',
-            isDefault: true
-        },
-        {
-            _id: '64492fd3d206f2a8b5f4298e',
-            name: 'В процессе',
-            color: '#d9cafe',
-            userId: '643c36ad8545463c883af7a8',
-            isDefault: false
-        }
-        ]
-    }
-}
-
 
 export const EditPageStoreDecorator: Decorator = (Story) => (
-    <StoreProvider initialState={state} asyncReducers={defaultAsyncReducers}>
+    <StoreProvider initialState={stateAllIn} asyncReducers={defaultAsyncReducers}>
         {Story()}
     </StoreProvider>
 )
