--- client/src/shared/ui/Calendar/ui/Calendar.stories.tsx
@@ -1,34 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-import 'app/styles/index.scss'
-import { WithColorsTheme } from 'shared/config/storybook/ThemeDecorator/WithColorsTheme'
-import { Calendar } from './Calendar'
-
-// More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
-const meta: Meta<typeof Calendar> = {
-    title: 'shared/Calendar',
-    component: Calendar,
-    argTypes: {
-    // backgroundColor: { control: 'color' },
-    },
-    tags: ['autodocs']
-}
-
-
-export default meta
-type Story = StoryObj<typeof Calendar>;
-
-// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
-
-export const Default: Story = {
-    args: {
-                                                                                                                                                          
-    },
-    decorators: [
-        WithColorsTheme
-    ],
-    parameters: {
-        docs: {
-            canvas: {sourceState: 'shown'}
-        },
-    }
-}
